requirements.txt
cachelib==0.10.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
Flask==2.2.3
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
Flask-Session==0.4.0
idna==3.4
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
mysqlclient==2.1.1
protobuf==3.20.3
pycparser==2.21
PyMySQL==1.0.3
requests==2.28.2
stripe==5.4.0
urllib3==1.26.15
Werkzeug==2.2.3
